Add dynamic humidity calibration from temp offset
With the physical limitations of the SuperSensor, we were able to
compensate for the temperature differential due to internal heating from
the ESP, but this compensation was not being taken into account by the
humidity sensor.

Add a compensation factor based on the offset to ensure humidity
properly tracks the filtered temperature value.